pyproject.toml
[tool.poetry]
name = "userbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
vkbottle = "^4.3.12"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
docstring-parser = "^0.16"
pydantic = "1.10.13"
tortoise-orm = "0.20.0"
aerich = "^0.7.2"
aiomysql = "^0.2.0"
pythonping = "^1.1.4"
matplotlib = "^3.9.1"
numpy = "^2.0.0"
qrcode = "^7.4.2"


[tool.aerich]
tortoise_orm = "config.tortoise_orm"
location = "./models/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
